fix: move dependency installation to install script

+++ install.sh
@@ -8,15 +8,22 @@ if [ "$EUID" -ne 0 ]
   exit
 fi
 
+echo "Installing Dependencies..."
+pacman -Sy
+pacman -S base base-devel xorg imlib2 --noconfirm
+
+echo "Installing Packages..."
 make -C ./suckless/chadwm clean install
 make -C ./suckless/slock clean install
 make -C ./suckless/slstatus clean install
 
-# get the user name
-USERNAME=$(logname)
-cp -R ./.dwm /home/$USERNAME/
-cp -R ./.config /home/$USERNAME/
-cp -R ./.wallpaper /home/$USERNAME/
-cp ./.xinitrc /home/$USERNAME/
-cp ./.zshrc /home/$USERNAME/
-cp ./.tmux.conf /home/$USERNAME/
+if [ -e /bin/logname ]; then
+  echo "Copying Configurations..."
+  USERNAME=`logname`
+  cp -R ./.dwm /home/$USERNAME/
+  cp -R ./.config /home/$USERNAME/
+  cp -R ./.wallpaper /home/$USERNAME/
+  cp ./.xinitrc /home/$USERNAME/
+  cp ./.zshrc /home/$USERNAME/
+  cp ./.tmux.conf /home/$USERNAME/
+fi
